Move some common files into common folder
