+++ hyperstyle/src/python/review/inspectors/common/tips.py
