+++ hyperstyle/src/python/review/inspectors/common/utils.py
