+++ hyperstyle/src/python/review/inspectors/eslint/eslint.py
@@ -4,10 +4,10 @@ from typing import Any, Dict, List
 from hyperstyle.src.python.review.common.file_system import new_temp_dir
 from hyperstyle.src.python.review.common.subprocess_runner import run_in_subprocess
 from hyperstyle.src.python.review.inspectors.base_inspector import BaseInspector
+from hyperstyle.src.python.review.inspectors.common.xml_parser import parse_xml_file_result
 from hyperstyle.src.python.review.inspectors.eslint.issue_types import ESLINT_CLASS_NAME_TO_ISSUE_TYPE
 from hyperstyle.src.python.review.inspectors.inspector_type import InspectorType
 from hyperstyle.src.python.review.inspectors.issue import BaseIssue, IssueDifficulty, IssueType
-from hyperstyle.src.python.review.inspectors.parsers.xml_parser import parse_xml_file_result
 
 PATH_ESLINT_CONFIG = Path(__file__).parent / '.eslintrc'
