+++ hyperstyle/src/python/review/inspectors/pyast/python_ast.py
@@ -7,11 +7,11 @@ from hyperstyle.src.python.review.common import language
 from hyperstyle.src.python.review.common.file_system import get_all_file_system_items
 from hyperstyle.src.python.review.common.language import Language
 from hyperstyle.src.python.review.inspectors.base_inspector import BaseInspector
+from hyperstyle.src.python.review.inspectors.common.tips import get_bool_expr_len_tip, get_func_len_tip
 from hyperstyle.src.python.review.inspectors.inspector_type import InspectorType
 from hyperstyle.src.python.review.inspectors.issue import (
     BaseIssue, BoolExprLenIssue, FuncLenIssue, IssueDifficulty, IssueType,
 )
-from hyperstyle.src.python.review.inspectors.tips import get_bool_expr_len_tip, get_func_len_tip
 
 BOOL_EXPR_LEN_ORIGIN_CLASS = 'C001'
 FUNC_LEN_ORIGIN_CLASS = 'C002'
