+++ hyperstyle/src/python/review/inspectors/radon/radon.py
@@ -4,7 +4,8 @@ from typing import Any, Dict, List
 
 from hyperstyle.src.python.review.common.subprocess_runner import run_in_subprocess
 from hyperstyle.src.python.review.inspectors.base_inspector import BaseInspector
-from hyperstyle.src.python.review.inspectors.common import convert_percentage_of_value_to_lack_of_value
+from hyperstyle.src.python.review.inspectors.common.tips import get_maintainability_index_tip
+from hyperstyle.src.python.review.inspectors.common.utils import convert_percentage_of_value_to_lack_of_value
 from hyperstyle.src.python.review.inspectors.inspector_type import InspectorType
 from hyperstyle.src.python.review.inspectors.issue import (
     BaseIssue,
@@ -13,7 +14,6 @@ from hyperstyle.src.python.review.inspectors.issue import (
     IssueType,
     MaintainabilityLackIssue,
 )
-from hyperstyle.src.python.review.inspectors.tips import get_maintainability_index_tip
 
 
 MAINTAINABILITY_ORIGIN_CLASS = 'RAD100'
